# wiki/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ops Agent API",
    "version": "0.1.0",
    "description": "Generated from registered Ops Agent routes. Middleware authentication, manually parsed request models and SSE media types are documented. WebSocket frames are described separately. Documentation viewer does not execute requests."
  },
  "paths": {
...
    "/api/jumpserver/instances": {
      "get": {
        "tags": [
          "jumpserver"
        ],
        "summary": "List Instances",
        "operationId": "list_instances_api_jumpserver_instances_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JumpServerInstanceView"
                  },
                  "type": "array",
                  "title": "Response List Instances Api Jumpserver Instances Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jumpserver"
        ],
        "summary": "Create Instance",
        "operationId": "create_instance_api_jumpserver_instances_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JumpServerInstanceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JumpServerInstanceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jumpserver/instances/{instance_id}": {
      "put": {
        "tags": [
          "jumpserver"
        ],
        "summary": "Update Instance",
        "operationId": "update_instance_api_jumpserver_instances__instance_id__put",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JumpServerInstanceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JumpServerInstanceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jumpserver/instances/{instance_id}/test": {
      "post": {
        "tags": [
          "jumpserver"
        ],
        "summary": "Test Instance",
        "operationId": "test_instance_api_jumpserver_instances__instance_id__test_post",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JumpServerOperationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jumpserver/instances/{instance_id}/sync": {
      "post": {
        "tags": [
          "jumpserver"
        ],
        "summary": "Sync Instance",
        "operationId": "sync_instance_api_jumpserver_instances__instance_id__sync_post",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JumpServerOperationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jumpserver/instances/{instance_id}/assets": {
      "get": {
        "tags": [
          "jumpserver"
        ],
        "summary": "List Instance Assets",
        "operationId": "list_instance_assets_api_jumpserver_instances__instance_id__assets_get",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JumpServerAssetBindingView"
                  },
                  "title": "Response List Instance Assets Api Jumpserver Instances  Instance Id  Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jumpserver/instances/{instance_id}/organizations": {
      "get": {
        "tags": [
          "jumpserver"
        ],
        "summary": "List Instance Organizations",
        "operationId": "list_instance_organizations_api_jumpserver_instances__instance_id__organizations_get",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jumpserver/assets/{binding_id}/account": {
      "patch": {
        "tags": [
          "jumpserver"
        ],
        "summary": "Select Asset Account",
        "operationId": "select_asset_account_api_jumpserver_assets__binding_id__account_patch",
        "parameters": [
          {
            "name": "binding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Binding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JumpServerAccountSelection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JumpServerAssetBindingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JumpServerAccountSelection": {
        "properties": {
          "account_ref": {
            "type": "string",
            "minLength": 1,
            "title": "Account Ref"
          }
        },
        "type": "object",
        "required": [
          "account_ref"
        ],
        "title": "JumpServerAccountSelection"
      },
      "JumpServerAccountView": {
        "properties": {
          "id": {
            "title": "Id"
          },
          "name": {
            "title": "Name"
          },
          "username": {
            "title": "Username"
          },
          "alias": {
            "title": "Alias"
          },
          "secret_type": {
            "title": "Secret Type"
          },
          "privileged": {
            "title": "Privileged"
          },
          "is_active": {
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "JumpServerAccountView"
      },
      "JumpServerAssetBindingView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "assetId": {
            "type": "integer",
            "title": "Assetid"
          },
          "externalAssetId": {
            "type": "string",
            "title": "Externalassetid"
          },
          "orgId": {
            "type": "string",
            "title": "Orgid",
            "default": ""
          },
          "orgName": {
            "type": "string",
            "title": "Orgname",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/JumpServerAccountView"
            },
            "type": "array",
            "title": "Accounts"
          },
          "accountRef": {
            "type": "string",
            "title": "Accountref"
          },
          "accountUsername": {
            "type": "string",
            "title": "Accountusername"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assetId",
          "externalAssetId",
          "name",
          "address",
          "platform",
          "category",
          "type",
          "accounts",
          "accountRef",
          "accountUsername",
          "active"
        ],
        "title": "JumpServerAssetBindingView"
      },
      "JumpServerInstanceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "auth_mode": {
            "type": "string",
            "enum": [
              "access_key",
              "ssh_gateway"
            ],
            "title": "Auth Mode",
            "default": "access_key"
          },
          "base_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Base Url"
          },
          "org_id": {
            "type": "string",
            "maxLength": 64,
            "title": "Org Id",
            "default": ""
          },
          "access_key_id": {
            "type": "string",
            "maxLength": 255,
            "title": "Access Key Id",
            "default": ""
          },
          "access_key_secret": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Key Secret"
          },
          "verify_tls": {
            "type": "boolean",
            "title": "Verify Tls",
            "default": true
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "base_url"
        ],
        "title": "JumpServerInstanceCreate"
      },
      "JumpServerInstanceUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "auth_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "access_key",
                  "ssh_gateway"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Mode"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "org_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Id"
          },
          "access_key_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Key Id"
          },
          "access_key_secret": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Key Secret"
          },
          "verify_tls": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify Tls"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "JumpServerInstanceUpdate"
      },
      "JumpServerInstanceView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "authMode": {
            "type": "string",
            "title": "Authmode"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "baseUrl": {
            "type": "string",
            "title": "Baseurl"
          },
          "orgId": {
            "type": "string",
            "title": "Orgid"
          },
          "accessKeyId": {
            "type": "string",
            "title": "Accesskeyid"
          },
          "accessKeySecretMasked": {
            "type": "string",
            "title": "Accesskeysecretmasked"
          },
          "verifyTls": {
            "type": "boolean",
            "title": "Verifytls"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "connectionStatus": {
            "type": "string",
            "title": "Connectionstatus"
          },
          "lastError": {
            "type": "string",
            "title": "Lasterror"
          },
          "lastSyncAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsyncat"
          },
          "assetCount": {
            "type": "integer",
            "title": "Assetcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "authMode",
          "name",
          "baseUrl",
          "orgId",
          "accessKeyId",
          "accessKeySecretMasked",
          "verifyTls",
          "enabled",
          "connectionStatus",
          "lastError",
          "lastSyncAt",
          "assetCount"
        ],
        "title": "JumpServerInstanceView"
      },
      "JumpServerOperationView": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "created": {
            "type": "integer",
            "title": "Created",
            "default": 0
          },
          "updated": {
            "type": "integer",
            "title": "Updated",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "JumpServerOperationView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}